pyproject.toml
[tool.poetry]
name = "async-postgres-docker"
version = "0.1.0"
description = "Base setup for a docker cluster with asynchronous postgres database and FastAPI framework with Alembic."
authors = ["ignacevau"]
readme = "README.md"
# packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
fastapi-sqlalchemy = "^0.2.1"
pydantic = "^1.10.2"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
fastapi = "^0.85.1"
gunicorn = "^20.1.0"
uvicorn = "^0.19.0"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"
requests = "^2.28.1"
httpx = "^0.23.0"
pytest-asyncio = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''


[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
profile = "black"
# line_length = 88
# indent = '    '
# multi_line_output = 3
# include_trailing_comma = true
# force_grid_wrap = 0


[tool.coverage]
# https://github.com/nedbat/coveragepy
    [tool.coverage.run]

    source = ["src"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]


[tool.pytest.ini_options]
pythonpath = [
  "."
]